Alterações em vistas para resolver erro de input do PrecoDia

+++ ViewModels/Camiao/CamiaoViewModel.cs
@@ -20,7 +20,7 @@ namespace Projeto_AutoMobile.ViewModels.Camiao
         [Required(ErrorMessage = "O preço por dia é obrigatório.")]
         [Range(0.01, double.MaxValue, ErrorMessage = "O preço por dia deve ser um valor positivo.")]
         [Display(Name = "Preço por Dia")]
-        public decimal PrecoDia { get; set; }
+        public double PrecoDia { get; set; }
 
         [Required(ErrorMessage = "O estado é obrigatório.")]
         public EstadoVeiculo Estado { get; set; }
