+++ ViewModels/Reserva/ReservaViewModel.cs
@@ -17,6 +17,6 @@ namespace Projeto_AutoMobile.ViewModels.Reserva
         [Required(ErrorMessage = "Selecione um veículo")]
         public int VeiculoId { get; set; }
 
-        public decimal PrecoEstimado { get; set; }
+        public double PrecoEstimado { get; set; }
     }
 }
